- Moved title and note to task meta data from task entity
