+++ server/src/entities/task/Task.options.ts
@@ -1,13 +1,3 @@
-export const TitleOptions: any = {
-    length: 150,
-    type: 'varchar',
-}
-
-export const NoteOptions: any = {
-    length: 2000,
-    type: 'varchar',
-    nullable: true,
-}
 
 export const DateOptions: any = {
     type: 'timestamptz',
