+++ server/src/entities/taskMetaData/TaskMetaData.options.ts
@@ -1,3 +1,13 @@
+export const TitleOptions: any = {
+    length: 150,
+    type: 'varchar',
+}
+
+export const NoteOptions: any = {
+    length: 2000,
+    type: 'varchar',
+    nullable: true,
+}
 
 export const StartDateOptions: any = {
     type: 'timestamptz',
