+++ server/src/typeDefs/taskMetaData.typeDef.ts
@@ -3,6 +3,8 @@ import { gql } from 'apollo-server'
 export const taskMetaDataType = gql`
     type TaskMetaData {
         id: String!,
+        title: String!,
+        note: String,
         startDate: GraphQLDateTime,
         endDate: GraphQLDateTime,
         rrule: String,
@@ -12,6 +14,8 @@ export const taskMetaDataType = gql`
     }
 
     input TaskMetaDataInput {
+        title: String!,
+        note: String,
         startDate: GraphQLDateTime,
         endDate: GraphQLDateTime,
         rrule: String,
